+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Fix Rails 8.2 compatibility (named store accessors module). ([@palkan][])
+
 ## 2.1.1 (2026-01-13)
 
 - Fix defining store attributes after schema loading. ([@palkan][])